\name{astrowrap_glm.diag}
\alias{astrowrap_glm.diag}
\title{Parameters for assessing GLM fits}
\description{
  Computes jackknife deviance residuals, standardized deviance residuals, standardized Pearson residuals, approximate Cook statistic, leverage, and estimated dispersion.
}

\usage{
astrowrap("glm.diag", ...)
}
\arguments{
  \item{...}{expressions evaluated in the context of the \code{\link[stats]{glm.diag}}}
}
\seealso{
\code{astrowrap_glm}, \code{astrowrap_summary}, \code{astrowrap_family}
}

